+++ examples/distributed/simple_sync_distributed.py
@@ -2,18 +2,19 @@
 """ Simple example of using leap_ec.distrib.synchronous
 
 """
+import os
+
 import multiprocessing.popen_spawn_posix  # Python 3.9 workaround for Dask.  See https://github.com/dask/distributed/issues/4168
 from distributed import Client
 import toolz
 
+from leap_ec import test_env_var
+from leap_ec import ops
 from leap_ec.decoder import IdentityDecoder
-import leap_ec.ops as ops
-
-from leap_ec.binary_rep.problems import MaxOnes
 from leap_ec.binary_rep.initializers import create_binary_sequence
 from leap_ec.binary_rep.ops import mutate_bitflip
-
-from leap_ec.distrib.individual import DistributedIndividual
+from leap_ec.binary_rep.problems import MaxOnes
+from leap_ec.distrib import DistributedIndividual
 from leap_ec.distrib import synchronous
 
 if __name__ == '__main__':
@@ -30,7 +31,15 @@ if __name__ == '__main__':
         # Scatter the initial parents to dask workers for evaluation
         parents = synchronous.eval_population(parents, client=client)
 
-        for current_generation in range(5):
+
+        # When running the test harness, just run for two generations
+        # (we use this to quickly ensure our examples don't get bitrot)
+        if os.environ.get(test_env_var, False) == 'True':
+            generations = 2
+        else:
+            generations = 5
+
+        for current_generation in range(generations):
             offspring = toolz.pipe(parents,
                                    ops.tournament_selection,
                                    ops.clone,