+++ examples/simple/int_rep.py
@@ -4,16 +4,16 @@ representation.
 We use a generational EA with binomial mutation of integer genes to minimize an
 integer version of the Langermann function.
 """
+import os
 import sys
 
 from matplotlib import pyplot as plt
 
+from leap_ec import Representation, test_env_var
+from leap_ec import ops, probe
 from leap_ec.algorithm import generational_ea
-from leap_ec.representation import Representation
-from leap_ec import ops
 from leap_ec.int_rep.initializers import create_int_vector
 from leap_ec.int_rep.ops import mutate_binomial
-from leap_ec import probe
 from leap_ec.real_rep.problems import LangermannProblem
 
 
@@ -26,9 +26,17 @@ if __name__ == '__main__':
     # we can also use it to evaluate integer-valued genomes.
     problem = LangermannProblem(maximize=False)
 
+
+    # When running the test harness, just run for two generations
+    # (we use this to quickly ensure our examples don't get bitrot)
+    if os.environ.get(test_env_var, False) == 'True':
+        generations = 2
+    else:
+        generations = 100
+    
     l = 2
     pop_size = 10
-    ea = generational_ea(generations=100,pop_size=pop_size,
+    ea = generational_ea(generations=generations,pop_size=pop_size,
                              problem=problem,  # Fitness function
 
                              # Representation